+++ app/app_components/loginModule/services/UserSettingsService.js
@@ -1,10 +1,14 @@
 'use strict';
 
 angular.module('loginModule')
-    .service('userSettingsService', ['$http', function ($http) {
+    .service('userSettingsService', ['$http', 'userService', function ($http, userService) {
 
         function persistUserSettings(user) {
-            return $http.put('/api/users/' + user.id + '/settings', user.userSettingsDto);
+            return $http.put('/api/users/' + user.id + '/settings', user.userSettingsDto).then(
+                function () {
+                    userService.clearCacheForGetCurrentUser();
+                    userService.getCurrentUser();
+                });
         }
 
         return {