+++ app/app_components/loginModule/services/userService.js
@@ -40,6 +40,8 @@ angular.module('loginModule')
         return {
             getCurrentUser: getCurrentUser,
             getUsersForLogin: getUsersForLogin,
-            logInUser: logInUser
+            logInUser: logInUser,
+
+            clearCacheForGetCurrentUser: clearCacheForGetCurrentUser
         };
     }]);